requirements.txt
Django==2.2.16
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary>=2.8
